Fix: update `.set_cores()` to handle edge cases
The previous version resulted would result in an error when the machine
contained only 1 or 2 cores. Now, when the machine has only 1 core, the
cluster creation will fail, with a message that not enough cores are
available. When the machine has 2 cores, all 2 cores are used. In all
other cases, the number of cores used is given by the number of cores
on the machine take one.